+++ package/kedro_viz/data_access/managers.py
@@ -12,7 +12,7 @@ from kedro.pipeline.node import Node as KedroNode
 from sqlalchemy.orm import sessionmaker
 
 from kedro_viz.constants import DEFAULT_REGISTERED_PIPELINE_ID, ROOT_MODULAR_PIPELINE_ID
-from kedro_viz.models.graph import (
+from kedro_viz.models.flowchart import (
     DataNode,
     GraphEdge,
     GraphNode,
