+++ package/kedro_viz/server.py
@@ -13,7 +13,7 @@ from kedro_viz.api.rest.responses import get_default_response
 from kedro_viz.data_access import DataAccessManager, data_access_manager
 from kedro_viz.database import create_db_engine
 from kedro_viz.integrations.kedro import data_loader as kedro_data_loader
-from kedro_viz.models.experiments_tracking import Base
+from kedro_viz.models.experiment_tracking import Base
 
 DEFAULT_HOST = "127.0.0.1"
 DEFAULT_PORT = 4141
