+++ package/tests/test_api/test_graphql/test_subscriptions.py
@@ -1,7 +1,7 @@
 import json
 
 from kedro_viz.api.graphql import schema
-from kedro_viz.models.experiments_tracking import RunModel
+from kedro_viz.models.experiment_tracking import RunModel
 
 
 class TestRunsAddedSubscription:
